crawler/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
meilisearch==0.31.0
lxml==4.9.3
urllib3==2.0.7
charset-normalizer==3.3.2
html5lib==1.1
Pillow==10.1.0
python-dotenv==1.0.0 